+++ notifications_utils/template.py
@@ -7,19 +7,10 @@ from flask import Markup
 
 from notifications_utils.columns import Columns
 from notifications_utils.renderers import HTMLEmail, SMSPreview, EmailPreview, LetterPreview
+from notifications_utils.field import Field
 
 
 class Template():
-
-    placeholder_pattern = re.compile(
-        '\(\('            # opening ((
-        '([^\(\)\?]+)'    # 1. name of placeholder, eg ‘registration number’
-        '(\?\?)?'         # 2. optional ??
-        '([^\)\(]*)'      # 3. optional text to display if the placeholder’s value is True
-        '\)\)'            # closing ))
-    )
-    placeholder_tag = "<span class='placeholder'>(({}{}))</span>"
-    optional_placeholder_tag = "<span class='placeholder-conditional'>(({}??</span>{}))"
 
     def __init__(
         self,
@@ -98,56 +89,31 @@ class Template():
         elif self.template_type == 'email' or not self.template_type:
             self._renderer = EmailPreview()
 
-    def get_match(self, match):
-        if match[1] and match[2]:
-            return match[0]
-        return match[0] + match[2]
-
-    def format_match(self, match):
-        if match.group(2) and match.group(3):
-            return self.optional_placeholder_tag.format(match.group(1), match.group(3))
-        return self.placeholder_tag.format(match.group(1), match.group(3))
-
-    def replace_match(self, match):
-        if match.group(2) and match.group(3):
-            return strip_html(match.group(3)) if str2bool(self.values.get(match.group(1))) else ''
-        return strip_html(self.values.get(match.group(1) + match.group(3)))
+    @property
+    def _raw_formatted(self):
+        return self.renderer(str(Field(self.content, {})))
 
     @property
-    def formatted(self):
-        return self.renderer(re.sub(
-            Template.placeholder_pattern, self.format_match, self.content
-        ))
+    def _raw_formatted_subject(self):
+        return str(Field(self.subject, {}))
 
     @property
     def formatted_subject(self):
-        return re.sub(
-            Template.placeholder_pattern, self.format_match, self.subject
-        )
+        return Markup(self._raw_formatted_subject)
 
     @property
-    def formatted_subject_as_markup(self):
-        return Markup(self.formatted_subject)
-
-    @property
-    def formatted_as_markup(self):
-        return Markup(self.formatted)
+    def formatted(self):
+        return Markup(self._raw_formatted)
 
     @property
     def placeholders(self):
-        return OrderedSet(
-            self.get_match(match) for match in re.findall(
-                Template.placeholder_pattern, (self.subject or '') + self.content
-            )
-        )
+        return Field(self.subject or '').placeholders | Field(self.content).placeholders
 
     @property
     def replaced(self):
         if self.missing_data:
             raise NeededByTemplateError(self.missing_data)
-        return self.renderer(re.sub(
-            Template.placeholder_pattern, self.replace_match, self.content
-        ))
+        return self.renderer(str(Field(self.content, self.values)))
 
     @property
     def replaced_content_count(self):
@@ -174,11 +140,7 @@ class Template():
     def replaced_subject(self):
         if self.missing_data:
             raise NeededByTemplateError(self.missing_data)
-        return re.sub(
-            Template.placeholder_pattern,
-            self.replace_match,
-            self.subject if self.subject else ""
-        )
+        return str(Field(self.subject, self.values))
 
     @property
     def missing_data(self):
@@ -235,13 +197,3 @@ class TemplateChange():
 
 def get_sms_fragment_count(character_count):
     return 1 if character_count <= 160 else math.ceil(float(character_count) / 153)
-
-
-def str2bool(value):
-    if not value:
-        return False
-    return str(value).lower() in ("yes", "y", "true", "t", "1", "include", "show")
-
-
-def strip_html(value):
-    return bleach.clean(value, tags=[], strip=True)
